+++ frontend/app/game/page.tsx
@@ -18,7 +18,7 @@ import { useGame } from "@/context/GameContext";
 import secureLocalStorage from "react-secure-storage";
 import { server_domain } from "../page";
 import axios from "axios";
-import { SpeedOption } from "./friend/optionselect/page";
+import { SpeedOption } from "@/type/type";
 
 const infomodalStyle = {
   position: "absolute" as "absolute",
